Partial refactor of save/load tests
