+++ alibi_detect/saving/_sklearn/saving.py
@@ -34,7 +34,7 @@ def save_model_config(model: BaseEstimator,
     filepath = base_path.joinpath(local_path)
     save_model(model, filepath=filepath, save_dir='model')
     cfg_model = {
-        'model_type': 'sklearn',
+        'flavour': 'sklearn',
         'src': local_path.joinpath('model')
     }
     return cfg_model
