+++ alibi_detect/saving/loading.py
@@ -1,4 +1,3 @@
-# TODO - Need to modularise torch and tensorflow imports and use. e.g. has_tensorflow and has_pytorch etc
 import logging
 import os
 from functools import partial
@@ -17,6 +16,8 @@ from alibi_detect.saving._tensorflow import load_detector_legacy, load_embedding
 from alibi_detect.saving._sklearn import load_model_sk
 from alibi_detect.saving.validate import validate_config
 from alibi_detect.base import Detector, ConfigurableDetector
+from alibi_detect.utils.frameworks import has_tensorflow, has_pytorch
+from alibi_detect.saving.schemas import SupportedModels_tf, SupportedModels_torch
 
 if TYPE_CHECKING:
     import tensorflow as tf
@@ -186,8 +187,7 @@ def _load_kernel_config(cfg: dict, backend: str = 'tensorflow') -> Callable:
     return kernel
 
 
-def _load_preprocess_config(cfg: dict,
-                            backend: Optional[str] = 'tensorflow') -> Optional[Callable]:
+def _load_preprocess_config(cfg: dict) -> Optional[Callable]:
     """
     This function builds a preprocess_fn from the preprocess dict in a detector config dict. The dict format is
     expected to match that generated by serialize_preprocess in alibi_detect.utils.saving (also see pydantic schema).
@@ -197,8 +197,6 @@ def _load_preprocess_config(cfg: dict,
     ----------
     cfg
         A preprocess_fn config dict. (see pydantic schemas).
-    backend
-        The backend.
 
     Returns
     -------
@@ -217,14 +215,20 @@ def _load_preprocess_config(cfg: dict,
             emb = kwargs.pop('embedding')  # embedding passed to preprocess_drift as `model` therefore remove
 
             # Backend specifics
-            if backend == 'tensorflow':
+            if has_tensorflow and isinstance(model, SupportedModels_tf):
                 model = prep_model_and_emb_tf(model, emb)
                 kwargs.pop('device')
-            elif backend == 'pytorch':  # TODO - once optional deps implemented
+            elif has_pytorch and isinstance(model, SupportedModels_torch):
                 raise NotImplementedError('Loading preprocess_fn for PyTorch not yet supported.')
 #               device = cfg['device'] # TODO - device should be set already - check
 #               kwargs.update({'model': kwargs['model'].to(device)})  # TODO - need .to(device) here?
 #               kwargs.update({'device': device})
+            elif model is None:
+                kwargs.pop('device')
+                model = emb
+            if model is None:
+                raise ValueError("A 'model'  and/or `embedding` must be specified when "
+                                 "preprocess_fn='preprocess_drift'")
             kwargs.update({'model': model})
         else:
             kwargs = cfg['kwargs']  # If generic callable, kwargs is cfg['kwargs']
@@ -251,7 +255,7 @@ def _load_model_config(cfg: dict) -> Callable:
     """
 
     # Load model
-    model_type = cfg['model_type']
+    flavour = cfg['flavour']
     src = cfg['src']
     custom_obj = cfg['custom_objects']
     layer = cfg['layer']
@@ -260,9 +264,9 @@ def _load_model_config(cfg: dict) -> Callable:
         raise FileNotFoundError("The `src` field is not a recognised directory. It should be a directory containing "
                                 "a compatible model.")
 
-    if model_type == 'tensorflow':
+    if flavour == 'tensorflow':
         model = load_model_tf(src, load_dir='.', custom_objects=custom_obj, layer=layer)
-    elif model_type == 'sklearn':
+    elif flavour == 'sklearn':
         model = load_model_sk(src, load_dir='.')
     else:
         raise NotImplementedError('Loading of PyTorch models not currently supported')
@@ -270,7 +274,7 @@ def _load_model_config(cfg: dict) -> Callable:
     return model
 
 
-def _load_embedding_config(cfg: dict, backend: str) -> Callable:  # TODO: Could type return more tightly
+def _load_embedding_config(cfg: dict) -> Callable:  # TODO: Could type return more tightly
     """
     Load a pre-trained text embedding from an embedding config dict.
 
@@ -278,8 +282,6 @@ def _load_embedding_config(cfg: dict, backend: str) -> Callable:  # TODO: Could 
     ----------
     cfg
         An embedding config dict. (see the pydantic schemas).
-    backend
-        The backend.
 
     Returns
     -------
@@ -288,7 +290,8 @@ def _load_embedding_config(cfg: dict, backend: str) -> Callable:  # TODO: Could 
     src = cfg['src']
     layers = cfg['layers']
     typ = cfg['type']
-    if backend == 'tensorflow':
+    flavour = cfg['flavour']
+    if flavour == 'tensorflow':
         emb = load_embedding_tf(src, embedding_type=typ, layers=layers)
     else:
         raise NotImplementedError('Loading of non-tensorflow embedding models not currently supported')
@@ -494,13 +497,13 @@ def resolve_config(cfg: dict, config_dir: Optional[Path]) -> dict:
             if key[-1] in ('model', 'proj'):
                 obj = _load_model_config(src)
             elif key[-1] == 'embedding':
-                obj = _load_embedding_config(src, backend)
+                obj = _load_embedding_config(src)
             elif key[-1] == 'tokenizer':
                 obj = _load_tokenizer_config(src)
             elif key[-1] == 'optimizer':
                 obj = _load_optimizer_config(src, backend)
             elif key[-1] == 'preprocess_fn':
-                obj = _load_preprocess_config(src, backend)
+                obj = _load_preprocess_config(src)
             elif key[-1] in ('kernel', 'x_kernel', 'c_kernel'):
                 obj = _load_kernel_config(src, backend)
